➖ Remove implicit dependency to hardhat-ethers in the deployment script.
