+++ hardhat.config.js
@@ -1,7 +1,6 @@
 require('dotenv').config()
-require('hardhat-deploy');
 require('@nomiclabs/hardhat-ethers')
-require("@nomicfoundation/hardhat-toolbox")
+require("@nomicfoundation/hardhat-chai-matchers");
 
 /** @type import('hardhat/config').HardhatUserConfig */
 module.exports = {
