+++ package.json
@@ -10,11 +10,10 @@
   "license": "MIT",
   "devDependencies": {
     "@glif/filecoin-address": "^2.0.1",
-    "@nomicfoundation/hardhat-toolbox": "^2.0.0",
-    "@nomiclabs/hardhat-ethers": "npm:hardhat-deploy-ethers@^0.3.0-beta.13",
+    "@nomicfoundation/hardhat-chai-matchers": "^1.0.4",
+    "@nomiclabs/hardhat-ethers": "^2.2.0",
     "dotenv": "^16.0.3",
     "ethers": "^5.7.1",
-    "hardhat": "^2.12.0",
-    "hardhat-deploy": "^0.11.18"
+    "hardhat": "^2.12.0"
   }
 }
