+++ test/SimpleCoin.js
@@ -1,19 +1,54 @@
-const { deployments, ethers } = require('hardhat')
+require('dotenv').config()
+const fa = require("@glif/filecoin-address");
+const { artifacts, ethers } = require('hardhat')
 const should = require('chai').should()
 
-let deployerF0Addr, deploymentTxHash, deploymentBlockHash, deploymentBlockNumber
+const DEPLOYER_PRIVATE_KEY = process.env.DEPLOYER_PRIVATE_KEY;
+let deployerF0Addr, deploymentTxHash, deploymentBlockHash, deploymentBlockNumber, simpleCoinAddress
 const otherAddress = '0xff000000000000000000000000000000deadbeef'
 
 describe('SimpleCoin', function () {
   it('Should successfully deploy', async function () {
-    await deployments.fixture(['SimpleCoin'])
-    const { linkedData, transactionHash } = await deployments.get('SimpleCoin')
-    deployerF0Addr = linkedData.f0Addr
-    deploymentTxHash = transactionHash
+    // use the deployer private key to compute the Filecoin f1 deployer address
+    // and get the right tx nonce
+    const deployer = new ethers.Wallet(DEPLOYER_PRIVATE_KEY);
+    const pubKey = ethers.utils.arrayify(deployer.publicKey);
+    const f1Addr = fa.newSecp256k1Address(pubKey).toString();
+
+    try {
+      // check that an actor has been deployed at the deployer address
+      let actorId = await ethers.provider.send("Filecoin.StateLookupID", [f1Addr, []]);
+      // format the deployer f0 address
+      actorId = ethers.utils.hexValue(Number(actorId.slice(1)))
+      deployerF0Addr =  ethers.utils.hexConcat(['0xff', ethers.utils.hexZeroPad(actorId, 19)])
+    } catch (e) {
+      console.error(`failed to resolve address ${f1Addr}. be sure to deploy an actor by sending FIL there`)
+      return
+    }
+
+    const maxPriorityFeePerGas = await ethers.provider.send("eth_maxPriorityFeePerGas", [])
+    const nonce = await ethers.provider.send("Filecoin.MpoolGetNonce", [f1Addr]);
+
+    const SimpleCoin = await ethers.getContractFactory("SimpleCoin");
+    const simpleCoin = await SimpleCoin.deploy( {
+      gasLimit: 1000000000,
+      maxPriorityFeePerGas,
+      nonce,
+    });
+
+    deploymentTxHash = simpleCoin.deployTransaction.hash
+    await simpleCoin.deployed();
   })
   it('Should access transaction details after it has been mined',
     async function () {
       const txByHash = await ethers.provider.getTransaction(deploymentTxHash)
+      const {
+        blockHash,
+        blockNumber,
+      } = txByHash
+      deploymentBlockHash = blockHash
+      deploymentBlockNumber = blockNumber
+
       txByHash.should.contain.keys(
         'blockHash',
         'blockNumber',
@@ -24,17 +59,13 @@ describe('SimpleCoin', function () {
       txByHash.from.should.be.a.properAddress
       txByHash.from.should.hexEqual(deployerF0Addr)
       should.not.exist(txByHash.to)
-      const {
-        blockHash,
-        blockNumber,
-      } = txByHash
-      deploymentBlockHash = blockHash
-      deploymentBlockNumber = blockNumber
     })
   it('Should access transaction receipt after it has been mined',
     async function () {
       const txReceipt = await ethers.provider.getTransactionReceipt(
         deploymentTxHash)
+      simpleCoinAddress = txReceipt.contractAddress
+
       txReceipt.should.contain.keys(
         'blockHash',
         'blockNumber',
@@ -94,10 +125,9 @@ describe('SimpleCoin', function () {
     })
 
     blockTxCountByHash.should.be.equal(blockTxCountByNumber)
-
   })
   it('Should interact with the contract using eth_call', async function () {
-    const SimpleCoin = await ethers.getContract('SimpleCoin')
+    const SimpleCoin = await ethers.getContractAt('SimpleCoin', simpleCoinAddress)
     const deployerBalance = await SimpleCoin.getBalance(deployerF0Addr)
     const receiverBalance = await SimpleCoin.getBalance(otherAddress)
 
@@ -106,13 +136,14 @@ describe('SimpleCoin', function () {
   })
   it('Should get the contract byte code at the deployed address',
     async function () {
-      const SimpleCoin = await ethers.getContract('SimpleCoin')
+      const SimpleCoin = await ethers.getContractAt('SimpleCoin', simpleCoinAddress)
       const code = await ethers.provider.getCode(SimpleCoin.address, "latest");
-      const { deployedBytecode } = await deployments.getArtifact('SimpleCoin')
+      const { deployedBytecode } = await artifacts.readArtifact("SimpleCoin")
+
       code.should.be.equal(deployedBytecode)
     })
   it('Should get storage using eth_getStorageAt', async function () {
-    const SimpleCoin = await ethers.getContract('SimpleCoin')
+    const SimpleCoin = await ethers.getContractAt('SimpleCoin', simpleCoinAddress)
 
     let key = ethers.utils.hexConcat([
       ethers.utils.hexZeroPad(deployerF0Addr, 32),
